requirements_complete.txt
absl-py==2.2.2
aiohttp==3.9.3
aiosignal==1.3.2
alembic==1.16.1
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
apify_client==1.9.2
apify_shared==1.3.2
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
Automat==25.4.16
av==13.1.0
bcrypt==4.3.0
billiard==4.2.1
black==24.2.0
blinker==1.9.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
check-requirements-txt==1.2.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloup==3.0.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.2
coverage==7.8.0
cryptography==45.0.2
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
ecdsa==0.19.1
executing==2.2.0
fastapi==0.115.12
fastapi-swagger==0.2.3
filelock==3.18.0
flake8==7.0.0
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
glcontext==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.2.2
google-generativeai==0.3.1
google-pasta==0.2.0
googleapis-common-protos==1.69.2
greenlet==3.0.1
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httptools==0.6.4
huggingface-hub==0.30.2
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
incremental==24.7.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isosurfaces==0.1.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.15.0
kiwisolver==1.4.8
kombu==5.5.3
lazy_loader==0.4
libclang==18.1.1
loguru==0.7.2
lxml==4.9.3
Mako==1.3.10
manim==0.19.0
manim-ml==0.0.24
ManimPango==0.6.0
mapbox_earcut==1.0.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.21.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.2.0
moderngl==5.12.0
moderngl-window==3.1.1
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.3.2
mypy==1.8.0
mypy-extensions==1.0.0
nano==0.10.0
narwhals==1.33.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.2.2
openai==1.6.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
pathspec==0.12.1
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.7
pluggy==1.5.0
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.1
Protego==0.4.0
proto-plus==1.26.1
protobuf==4.25.6
proxy-manager==0.0.6
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydantic-settings==2.9.1
pydeck==0.9.1
PyDispatcher==2.0.7
pydub==0.25.1
pyee==11.0.1
pyfiglet==1.0.2
pyflakes==3.2.0
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PythonTurtle==0.3.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
queuelib==1.8.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
screeninfo==0.8.1
seaborn==0.13.2
selenium==4.33.0
sentry-sdk==1.40.6
service-identity==24.2.0
setuptools==65.5.0
shapely==2.1.1
six==1.17.0
skia-pathops==0.8.0.post2
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
srt==3.5.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.32.0
svgelements==1.9.6
sympy==1.14.0
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.24
tldextract==5.3.0
tokenizers==0.15.2
toml==0.10.2
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
transformers==4.37.2
trio==0.30.0
trio-websocket==0.12.2
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
undetected-chromedriver==3.5.5
urllib3==2.4.0
user-agents==2.2.0
uvicorn==0.34.3
vine==5.1.0
w3lib==2.3.1
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.19.0
yt-dlp==2025.5.22
zope.interface==7.2
